move enterprise/internal/codeintel/context to internal/codeintel/codycontext
