+++ enterprise/cmd/worker/internal/embeddings/repo/BUILD.bazel
@@ -17,7 +17,7 @@ go_library(
         "//cmd/worker/job",
         "//cmd/worker/shared/init/db",
         "//enterprise/cmd/worker/shared/init/codeintel",
-        "//enterprise/internal/codeintel/context",
+        "//internal/codeintel/codycontext",
         "//enterprise/internal/database",
         "//internal/embeddings",
         "//internal/embeddings/background/repo",
