+++ internal/codeintel/codycontext/testdata/TestSplitIntoEmbeddableChunks.golden
