+++ internal/embeddings/embed/BUILD.bazel
@@ -11,7 +11,7 @@ go_library(
     importpath = "github.com/sourcegraph/sourcegraph/internal/embeddings/embed",
     visibility = ["//enterprise:__subpackages__"],
     deps = [
-        "//enterprise/internal/codeintel/context",
+        "//internal/codeintel/codycontext",
         "//internal/embeddings",
         "//internal/embeddings/background/repo",
         "//internal/embeddings/embed/client",
@@ -38,7 +38,7 @@ go_test(
     ],
     embed = [":embed"],
     deps = [
-        "//enterprise/internal/codeintel/context",
+        "//internal/codeintel/codycontext",
         "//internal/embeddings/background/repo",
         "//internal/embeddings/embed/client",
         "//internal/api",
