+++ internal/embeddings/embed/iface.go
@@ -3,7 +3,7 @@ package embed
 import (
 	"context"
 
-	codeintelContext "github.com/sourcegraph/sourcegraph/enterprise/internal/codeintel/context"
+	codeintelContext "github.com/sourcegraph/sourcegraph/internal/codeintel/codycontext"
 )
 
 type ContextService interface {
