+++ internal/embeddings/embed/mocks_test.go
@@ -10,7 +10,7 @@ import (
 	"context"
 	"sync"
 
-	context1 "github.com/sourcegraph/sourcegraph/enterprise/internal/codeintel/context"
+	context1 "github.com/sourcegraph/sourcegraph/internal/codeintel/codycontext"
 )
 
 // MockContextService is a mock implementation of the ContextService
